basically finish exchange voucher

+++ backend/user-service/internal/utils/jwt_utils.go
@@ -8,7 +8,7 @@ import (
 	"github.com/golang-jwt/jwt/v5"
 )
 
-func CreateToken(ID int16, username string) (string, error) {
+func CreateToken(ID int64, username string) (string, error) {
 	token := jwt.NewWithClaims(jwt.SigningMethodHS256,
 		jwt.MapClaims{
 			"username": username,
